Change service schema to support non-uniform waypoint setup in multi-cluster
Currently multi-cluster support in ambient mode assumes that waypoints
in all the clusters are configured the same way (e.g., the waypoint
service is named the same across the clusters and the same service will
have or not have a waypoint in all the clusters simultaneously).

It's somewhat undersirable, because we want to be able to configure
clusters somewhat independently of each other. This change updates pilot
to keep a mapping of clusters to waypoint in the service in addition to
a single waypoint field.

The idea is to give ztunnel this updated structure and ztunnel, roughtly
speaking, will first select a cluster it wants to send the request to
(complying with the load balancing policies, i.e., it should not select
a cluster on a remote network if the policy wants the traffic to stay
locally) and then select the backend in that cluster. And if the
selected cluster has a waypoint, then we will pick a waypoint backend.

NOTE: It could happen that there is no healthy backends in a cluster
ztunnel selects, in this case, depending on the laod balancing policy
we would want to consider another cluster, so it's not exactly as simple
as first selecting the cluster and then selecting a backend in that
cluster, but it should give you a general idea of what is going on.

NOTE: I've also attached some weight to clusters and currently the
weight is always 1. The idea is to pick the cluster of the available
clusters with a likelyhood proportional to the weight, but that's a bit
of a stretch at this point (e.g., how do users will even set the weight?,
do we even want pilot to popultate the weight or distribute that
information in some other fashion? - all-in-all, there are multiple open
questions there, so in the final version I'll probably drop the weight).

Signed-off-by: Mikhail Krinkin <mkrinkin@microsoft.com>